Splits backprop.py in two files, one of which can be converted to C
PiperOrigin-RevId: 171165855
